Добавил Fresco

+++ app/src/main/java/com/github/alexandrgrebenkin/weatherapp/data/rest/openweathermap/entities/DailyRestModel.java
@@ -31,7 +31,7 @@ public class DailyRestModel {
     @SerializedName("clouds")
     public int clouds;
     @SerializedName("pop")
-    public int pop;
+    public float pop;
     @SerializedName("rain")
     public float rain;
     @SerializedName("uvi")
